Replace bandit with ruff
ruff has reimplemented the bandit rules[1], so we can use that as a
better-integrated tool with the rest of our stack. We enable all the
bandit rules and selectively disable some across the codebase and some
in just tests where they don't make sense (e.g. flagging use of
`assert` or using insecure crypto).

For bonus points we can get rid of the GitPython dependency, which has a
history of (non-exploitable in our context) security issues.

[1] Per https://github.com/astral-sh/ruff/issues/1646 they've
    implemented nearly all of them, and the remaining ones aren't that
    important IMO.