Run `make lint` job in GitHub Actions
This is a ground-up rewrite of the "lint" job instead of porting it from
CircleCI. It will be much faster since it doesn't try to spin up a
Docker container and work through our various unnecessary caching
layers. It does literally the bare minimum to run lints.

The CircleCI lint job needs to remain for now because it's super
intertwined with everything else.